Recent updates, add xmldiff and findupthumbs

+++ Wallpapers/clean.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
 
 rm -f *.txt
-for i in */*.txt
+for i in */*.txt */*/*.txt */login.? */*/login.?
 do
+    [ "$i" = "*/*.txt" ] && continue
+    [ "$i" = "*/*/*.txt" ] && continue
+    [ "$i" = "*/login.?" ] && continue
+    [ "$i" = "*/*/login.?" ] && continue
     basnam=`basename "$i"`
     [ "$basnam" = "downloaded.txt" ] || rm -f "$i"
 done
